Added control template (Need help with getting bindings from BaseVm)
