+++ CheapShark Project/csApiApp-master/csApiApp/csApiApp/Mvvm/View/BasePage.xaml
@@ -1,12 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="csApiApp.Mvvm.View.BasePage">
-    <ContentPage.Content>
-        <StackLayout>
-            <Label Text="Welcome to Xamarin.Forms!"
-                VerticalOptions="CenterAndExpand" 
-                HorizontalOptions="CenterAndExpand" />
-        </StackLayout>
-    </ContentPage.Content>
+             xmlns:z="clr-namespace:FunctionZero.zBind.z;assembly=FunctionZero.zBind"
+             x:Class="csApiApp.Mvvm.View.BasePage"
+             ControlTemplate="{StaticResource BaseTemplate}">
 </ContentPage>
