+++ CheapShark Project/csApiApp-master/csApiApp/csApiApp/Mvvm/View/HomePage.xaml.cs
@@ -4,7 +4,7 @@ using Xamarin.Forms.Xaml;
 namespace csApiApp.Mvvm.View
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
-    public partial class HomePage : ContentPage
+    public partial class HomePage : BasePage
     {
         public HomePage()
         {
